Move example component selection into container
This moves the selection of the example component based on the
example's type (e.g., `<HandlebarsExample>`) into the
`<ExampleContainer>` component.
