+++ lib/examples.js
@@ -54,9 +54,8 @@ function renderExamples(examples, intl) {
 
     return examples.reduce(function (hash, example) {
         hash[example.name] = component.render('example-container', {
-            component: component.require(example.type + '-example'),
-            example  : example,
-            intl     : intl
+            example: example,
+            intl   : intl
         });
 
         return hash;
